cont

+++ src/picoc/cstdlib/unistd.c
@@ -501,9 +501,9 @@ const char UnistdDefs[] = "\
 typedef int uid_t; \
 typedef int gid_t; \
 typedef int pid_t; \
-typedef int off_t; \
-typedef int size_t; \
-typedef int ssize_t; \
+typedef long off_t; \
+typedef unsigned long size_t; \
+typedef long ssize_t; \
 typedef int useconds_t;\
 typedef int intptr_t;\
 ";
